fix(revenue): keep recovery seconds contract integer

+++ app/Domain/Commerce/Services/PendingCheckoutRecoveryService.php
@@ -91,7 +91,7 @@ final class PendingCheckoutRecoveryService
             'payment_recovery_eligible' => true,
             'payment_expires_at' => $expiresAt?->toIso8601String(),
             'payment_recovery_seconds_remaining' => $expiresAt
-                ? max(0, now()->diffInSeconds($expiresAt, false))
+                ? (int) max(0, now()->diffInSeconds($expiresAt, false))
                 : 0,
         ];
     }
